refactor(review): clear remaining low thermo findings
Use ledger get_record for convergence state lookup, share verdict
vocabulary via evals/verdict_vocab, promote withdrawn_fingerprints_for_state,
wire _path_from_record in deferred promotion, and move recall corpus
fixtures into convergence_benchmark.
